aula_05_0/requirements.txt
numpy
PyOpenGL
PyOpenGL-accelerate
pygame
opencv-python
matplotlib